Part 6 refactor

+++ include/part_1_foundations/4_divide_and_conquer/divide_and_conquer.h
@@ -21,12 +21,12 @@ void matrix_add_recursive(
 // CHAPTER 4.2: STRASSEN ALGORITHM
 // ============================================================================
 
-void cleanup_strassen_matrices(double** A11, double** A12, double** A21, double** A22,
-                            double** B11, double** B12, double** B21, double** B22,
-                            double** S1, double** S2, double** S3, double** S4, double** S5,
-                            double** S6, double** S7, double** S8, double** S9, double** S10,
-                            double** P1, double** P2, double** P3, double** P4, double** P5,
-                            double** P6, double** P7, int newSize);
-void strassen_algorithm(double** A, double** B, double** C, int size);
+void cleanup_strassen_matrices(int** A11, int** A12, int** A21, int** A22,
+                            int** B11, int** B12, int** B21, int** B22,
+                            int** S1, int** S2, int** S3, int** S4, int** S5,
+                            int** S6, int** S7, int** S8, int** S9, int** S10,
+                            int** P1, int** P2, int** P3, int** P4, int** P5,
+                            int** P6, int** P7, int newSize);
+void strassen_algorithm(int** A, int** B, int** C, int size);
 
 #endif // DIVIDE_AND_CONQUER_H
